add utils.js
